separate navbar into more files and make mobile responsive navbar
