+++ components/footer.jsx
@@ -1,10 +1,13 @@
-import React from "react";
+import MobileNavbarIcons from "./mobileNavbarIcons";
 
 const Footer = () => {
   return (
-    <footer className="bg-gray-800 text-white p-4 text-center mt-52">
-      <p>&copy; 2024 SnapChef. All rights reserved.</p>
-    </footer>
+    <>
+      <footer className="bg-gray-800 text-white p-4 text-center mt-52 sm:block hidden">
+        <p>&copy; 2024 SnapChef. All rights reserved.</p>
+      </footer>
+      <MobileNavbarIcons />
+    </>
   );
 };
