+++ components/navbar.jsx
@@ -3,51 +3,29 @@ import Image from "next/image";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { useState, useEffect } from "react";
-import SignInButton from "../app/signin/components/signInButton";
 import SnapChefLogo from "../assets/LogoDesign1.svg";
 import Search from "../assets/icons/search.svg";
-import Home from "../assets/icons/home.svg";
-import HomeFill from "../assets/icons/homeFill.svg";
-import Bell from "../assets/icons/bell.svg";
-import BellFill from "../assets/icons/bellFill.svg";
-import Post from "../assets/icons/post.svg";
-import PostFill from "../assets/icons/postFill.svg";
-import Fav from "../assets/icons/favorites.svg";
-import FavFill from "../assets/icons/favoritesFill.svg";
-import Profile from "../assets/icons/profile.svg";
-import ProfileFill from "../assets/icons/profileFill.svg";
 import Autosuggest from "react-autosuggest";
 import Fuse from "fuse.js";
 import { fetchRecipeNames } from "@/actions/fetches";
-import Notifications from "./notifications";
+import NavbarIcons from "./navbarIcons";
 
 const fuseSettings = {
   keys: ["recipe_name"], // key to search
   threshold: 0.4, // set the threshold for fuzzy search
 };
 
-import { useSession } from "next-auth/react";
-
 function Navbar() {
   const pathname = usePathname();
   const [activePath, setActivePath] = useState("");
-  const [showNotifications, setShowNotifications] = useState(false); // State to control the visibility of Notifications component
   const [searchTerm, setSearchTerm] = useState("");
   const [searchResults, setSearchResults] = useState([]);
   const [suggestions, setSuggestions] = useState([]);
-
-  const { data: session } = useSession();
-
-  const handleBellClick = () => {
-    setShowNotifications(!showNotifications); // Toggle the visibility of Notifications component
-  };
 
   useEffect(() => {
     // Update active path whenever the route changes
     setActivePath(pathname);
   }, [pathname]);
-
-  const isLinkActive = (path) => path === activePath;
 
   // sets URL to fetch recipes based on search term
   const handleSubmit = (e) => {
@@ -137,7 +115,7 @@ function Navbar() {
   };
 
   return (
-    <nav className="flex items-center justify-between p-4 relative mr-2">
+    <nav className="flex items-center justify-between p-4 relative mr-10">
       {/* Use the larger SnapChef.svg logo */}
       <Link href="/">
         <Image
@@ -174,68 +152,7 @@ function Navbar() {
           </div>
         </form>
       )}
-
-      {session ? (
-        <ul className="flex gap-8 list-none">
-          <Link href="/home">
-            <Image
-              src={isLinkActive("/home") ? HomeFill : Home}
-              alt="Home"
-              className="nav-icon cursor-pointer"
-              width={40}
-              height={40}
-            />
-          </Link>
-          <div className="nav-icon cursor-pointer" onClick={handleBellClick}>
-            <Image
-              src={showNotifications ? BellFill : Bell}
-              alt="Bell"
-              width={40}
-              height={40}
-            />
-          </div>
-          <Link href="/create">
-            <Image
-              src={isLinkActive("/create") ? PostFill : Post}
-              alt="Post"
-              className="nav-icon cursor-pointer"
-              width={40}
-              height={40}
-            />
-          </Link>
-          <Link href={`/favorites/${session?.user?.name}`}>
-            <Image
-              src={
-                isLinkActive(`/favorites/${session?.user?.name}`)
-                  ? FavFill
-                  : Fav
-              }
-              alt="Home"
-              className="nav-icon cursor-pointer"
-              width={40}
-              height={40}
-            />
-          </Link>
-          <Link href={`/profile/${session?.user?.name}`}>
-            <Image
-              src={
-                isLinkActive(`/profile/${session?.user?.name}`)
-                  ? ProfileFill
-                  : Profile
-              }
-              alt="Home"
-              className="nav-icon cursor-pointer"
-              width={40}
-              height={40}
-            />
-          </Link>
-        </ul>
-      ) : (
-        <SignInButton />
-      )}
-      {showNotifications && (
-        <Notifications setShowNotifications={setShowNotifications} />
-      )}
+      <NavbarIcons />
     </nav>
   );
 }
